requirements.txt
dash
dash-bootstrap-components
plotly
flask
requests
opencv-python
numpy
pillow
scikit-image
scikit-learn
pandas